requirements.txt
pandas==2.1.1
numpy==1.26.4
pymongo==4.7.1
python-dateutil==2.9.0
python-dotenv